--- a/tests/test_time_averaged_local_vs_traditional.py
+++ b/tests/test_time_averaged_local_vs_traditional.py
@@ -1,5 +1,4 @@
 import os
-import numpy as np
 import pandas as pd
 import torch