… OPTIMIZATION_NOTES.md
… thoi/measures/gaussian_copula.py